requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
_py-xgboost-mutex=2.0=gpu_0
alsa-lib=1.2.8=h166bdaf_0
attr=2.5.1=h166bdaf_1
blas=1.0=mkl
brotli=1.0.9=h166bdaf_9
brotli-bin=1.0.9=h166bdaf_9
brotli-python=1.0.9=py38hfa26641_9
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.5.7=hbcca054_0
certifi=2023.5.7=pyhd8ed1ab_0
charset-normalizer=3.2.0=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
combo=0.1.3=pypi_0
cuda-version=11.3=hbc958af_2
cudatoolkit=11.3.1=hb98b00a_12
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.6=h5008d03_3
expat=2.5.0=hcb278e6_1
ffmpeg=4.3=hf484d3e_0
fftw=3.3.10=nompi_hc118613_108
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.40.0=py38h01eb140_0
freetype=2.12.1=hca18f0e_1
gettext=0.21.1=h27087fc_0
glib=2.76.4=hfc55251_0
glib-tools=2.76.4=hfc55251_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
gst-plugins-base=1.21.3=h4243ec0_1
gstreamer=1.21.3=h25f0c4b_1
gstreamer-orc=0.4.34=hd590300_0
icu=70.1=h27087fc_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.8.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jack=1.9.22=h11f4161_0
joblib=1.3.0=pyhd8ed1ab_1
jpeg=9e=h0b41bf4_3
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py38h43d8883_1
krb5=1.20.1=h81ceb04_0
lame=3.100=h166bdaf_1003
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libblas=3.9.0=12_linux64_mkl
libbrotlicommon=1.0.9=h166bdaf_9
libbrotlidec=1.0.9=h166bdaf_9
libbrotlienc=1.0.9=h166bdaf_9
libcap=2.67=he9d0100_0
libcblas=3.9.0=12_linux64_mkl
libclang=15.0.7=default_h7634d5b_2
libclang13=15.0.7=default_h9986a30_2
libcups=2.3.3=h36d4200_3
libdb=6.2.32=h9c3ff4c_0
libdeflate=1.17=h0b41bf4_0
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.10=h28343ad_4
libexpat=2.5.0=hcb278e6_1
libffi=3.4.2=h7f98852_5
libflac=1.4.3=h59595ed_0
libgcc-ng=13.1.0=he5830b7_0
libgcrypt=1.10.1=h166bdaf_0
libgfortran-ng=13.1.0=h69a702a_0
libgfortran5=13.1.0=h15d22d2_0
libglib=2.76.4=hebfc3b9_0
libgomp=13.1.0=he5830b7_0
libgpg-error=1.47=h71f35ed_0
libiconv=1.17=h166bdaf_0
liblapack=3.9.0=12_linux64_mkl
libllvm15=15.0.7=hadd5161_1
libnsl=2.0.0=h7f98852_0
libogg=1.3.4=h7f98852_1
libopus=1.3.1=h7f98852_1
libpng=1.6.39=h753d276_0
libpq=15.3=hbcd7760_1
libsndfile=1.2.0=hb75c966_0
libsqlite=3.42.0=h2797004_0
libstdcxx-ng=13.1.0=hfd8a6a1_0
libsystemd0=253=h8c4010b_1
libtiff=4.5.0=h6adf6a1_2
libtool=2.4.7=h27087fc_0
libudev1=253=h0b41bf4_1
libuuid=2.38.1=h0b41bf4_0
libuv=1.44.2=h166bdaf_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.3.1=hd590300_0
libxcb=1.13=h7f98852_1004
libxgboost=1.7.4=cuda112h075c576_1
libxkbcommon=1.5.0=h79f4944_0
libxml2=2.10.3=hca2bb57_4
libzlib=1.2.13=hd590300_5
llvmlite=0.40.1=pypi_0
lz4-c=1.9.4=hcb278e6_0
matplotlib=3.5.1=py38h578d9bd_0
matplotlib-base=3.5.1=py38hf4fb855_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h95df7f1_0
mkl_fft=1.3.1=py38h8666266_1
mkl_random=1.2.2=py38h1abd341_0
mpg123=1.31.3=hcb278e6_0
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=8.0.33=hf1915f5_1
mysql-libs=8.0.33=hca2cd23_1
nccl=2.18.3.1=h12f7317_0
ncurses=6.4=hcb278e6_0
nettle=3.6=he412f7d_0
nspr=4.35=h27087fc_0
nss=3.89=he45b914_0
numba=0.57.1=pypi_0
numpy=1.24.3=py38h14f4228_0
numpy-base=1.24.3=py38h31eccc5_0
openh264=2.1.1=h780b84a_0
openjpeg=2.5.0=hfec8fc6_2
openssl=3.1.1=hd590300_1
packaging=23.1=pyhd8ed1ab_0
pandas=1.4.1=py38h43a58ef_0
pcre2=10.40=hc3806b6_0
pillow=9.4.0=py38hde6dc18_1
pip=23.1.2=pyhd8ed1ab_0
platformdirs=3.8.1=pyhd8ed1ab_0
ply=3.11=py_1
pooch=1.7.0=pyha770c72_3
psutil=5.9.5=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
pulseaudio=16.1=hcb278e6_3
pulseaudio-client=16.1=h5195f5e_3
pulseaudio-daemon=16.1=ha8d29e2_3
py-xgboost=1.7.4=cuda112py38hded6998_1
pyod=1.1.0=pypi_0
pyparsing=3.1.0=pyhd8ed1ab_0
pyqt=5.15.7=py38ha0d8c90_3
pyqt5-sip=12.11.0=py38h8dc9893_3
pysocks=1.7.1=pyha2e5f31_6
python=3.8.17=he550d4f_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=3_cp38
pytorch=1.10.1=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2023.3=pyhd8ed1ab_0
pyyaml=6.0=py38h0a891b7_5
qt-main=5.15.6=hf6cd601_5
readline=8.2=h8228510_1
requests=2.31.0=pyhd8ed1ab_0
scikit-learn=1.0.2=py38h1561384_0
scipy=1.10.1=py38h59b608b_3
setuptools=68.0.0=pyhd8ed1ab_0
sip=6.7.9=py38h17151c0_0
six=1.16.0=pyh6c4a22f_0
suod=0.0.9=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h27826a3_0
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=pyhd8ed1ab_0
torchaudio=0.10.1=py38_cu113
torchvision=0.11.2=py38_cu113
tornado=6.3.2=py38h01eb140_0
tqdm=4.63.0=pyhd8ed1ab_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
unicodedata2=15.0.0=py38h0a891b7_0
urllib3=2.0.3=pyhd8ed1ab_1
wheel=0.40.0=pyhd8ed1ab_0
xcb-util=0.4.0=h166bdaf_0
xcb-util-image=0.4.0=h166bdaf_0
xcb-util-keysyms=0.4.0=h166bdaf_0
xcb-util-renderutil=0.3.9=h166bdaf_0
xcb-util-wm=0.4.1=h166bdaf_0
xgboost=1.7.4=cuda112py38he5377ca_1
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
zipp=3.16.2=pypi_0
zlib=1.2.13=hd590300_5
zstd=1.5.2=hfc55251_7
